Create invitation controller
