+++ packages/editor-ui/src/views/SignupView.vue
@@ -83,14 +83,8 @@ export default defineComponent({
 		};
 	},
 	async mounted() {
-		const inviterId =
-			!this.$route.query.inviterId || typeof this.$route.query.inviterId !== 'string'
-				? null
-				: this.$route.query.inviterId;
-		const inviteeId =
-			!this.$route.query.inviteeId || typeof this.$route.query.inviteeId !== 'string'
-				? null
-				: this.$route.query.inviteeId;
+		const inviterId = this.getQueryParameter('inviterId');
+		const inviteeId = this.getQueryParameter('inviteeId');
 		try {
 			if (!inviterId || !inviteeId) {
 				throw new Error(this.$locale.baseText('auth.signup.missingTokenError'));
@@ -129,7 +123,7 @@ export default defineComponent({
 
 			try {
 				this.loading = true;
-				await this.usersStore.signup({
+				await this.usersStore.acceptInvitation({
 					...values,
 					inviterId: this.inviterId,
 					inviteeId: this.inviteeId,
@@ -152,6 +146,11 @@ export default defineComponent({
 				this.showError(error, this.$locale.baseText('auth.signup.setupYourAccountError'));
 			}
 			this.loading = false;
+		},
+		getQueryParameter(key: 'inviterId' | 'inviteeId'): string | null {
+			return !this.$route.query[key] || typeof this.$route.query[key] !== 'string'
+				? null
+				: (this.$route.query[key] as string);
 		},
 	},
 });
